Adjust for staging/production environment, extend supermarket cookbook.
Extends the existing supermarket cookbook for a staging/production environment while renaming the vagrant recipe to vagrant (which still may need adjustments but vagrant doesn't appear to be working for anyone now as is so it can be dealt with in the near future). It also makes a few adjustments to the codebase that are necessary for staging/production deployment. A Capistrano recipe and deployment documentation is also included.