Replace Devise authentication with ChefOAuth2.
This replaces Devise (email/password) authentication with Omniauth
authentication. Supermarket uses oc-id, a Chef service, as the provider for
authentication. This means that for a user to sign up and/or sign in to
Supermarket they must create a Chef Server account, as they currently do. This
however removes the duplication of records and simplifies the communication
between the community site and Chef Server for authentication when sharing
cookbooks.

Notes:

* Remove omniauth-twitter because it is no longer being used
* Store chef_oauth2 public key on Users.
* Configures the cookbook