Improve Supermarket cookbook
The .env template is pretty reliant on much of what happens in
deploy_revision. It was moved outside of deploy_revision so that the
.env template could be changed without having to have a different commit
hash available to deploy. This adds action to deploy_revision with a
data bag configurable attribute deploy_action that could be used to
force a deploy and alter .env.

Rather than checking for the contents of the page this changes it so it
just looks for a HTML5 document so the app cookbook specs don't have to
change everytime the apps content changes.

Alters the postgres recipe so the supermarket user only has login
permissions and is granted access to the database created for
Supermarket. Since SUPERUSER permissions are required to enable
extensions this moves extension creation into the postgres recipe and
removes it from the application schema/migrations.

The application recipe now generates a database.yml for use in
production on the fly based on the recipes attributes for postgres
database, username and pool size.