ensure pg gem installs for app
The application deployment does a bundle install. The bundle includes
the `pg` gem. The `pg` gem has C extensions. The application may not
necessarily run on a system runnning postgresql (and probably won't).
So, make sure that the ruby recipe includes all the right packages to
build C extensions in `bundle install`.