add rake db:seed to deploy resource
This adds the rake db:seed command to the deploy resource to ensure
the database gets populated with the ICLA/CCLA. This is done during
the deployment after restart, as the database will be set up, and we
want to ensure that this happens. The rake task itself is idempotent,
so re-running it should have no side effects on the database.

This also ensures the proper version of postgresql is used based on
the platform. Due to path locations, the `db:migrate` and `db:seed`
commands were failing at different points. This adds an attribute for
the postgres version. By default it will be 9.1, and can be changed by
the end user if necessary.