allow http domains specified as an array
The allow_http_domain attribute should be set as an attribute. This is
used by CHEF to maintain compatibility with the cookbooks.opscode.com
and api.berkshelf.com domains used in Berkshelf 2 and 3.

Previously, setting this as a string worked because we only had
api.berkshelf.com, but now we need to support the legacy cookbooks
site via HTTP.