use attribute for unicorn worker count in runit
The change to the .env file to add WEB_CONCURRENCY was not getting
picked up - the app was still only using 3 workers. Setting it in the
run script for the runit service definitely works.

This is set as an attribute so we don't have to load the data bag item
in the runit recipe, and we can just consume the data. It also means
we can just use a number of workers equal to the number of CPUs in the
system (using the `node['cpu']['total']` attribute).