Decouple the ruby recipe
The `supermarket::_ruby` recipe uses the `rubies` cookbook, which ends
up compiling from source the desired version of Ruby. This is highly
undesirable in our production supermarket environment as we do NOT
want to be in the business of unnecessarily compiling Ruby on
machines. It adds several minutes to the build time of new instances
and it is obtuse. The `_ruby` previously performed a package install
from the "Brightbox" repositories. The `supermarket::default` recipe
still includes `supermarket::_ruby`, so a "full stack" installation
will continue to work (e.g., test kitchen, normal standalone
supermarket installs, etc).

While we're making progress on the omnibus supermarket, it's not done
yet. This change makes our staging and production environments not
require a custom compiled Ruby, as we can use our own `_ruby` recipe.
Also, this ensures maximum flexibility for others who may wish to
install Ruby from local package repositories, for example.