Remove git references for dependent cookbooks
Using the git references for cookbook dependencies means that running
test kitchen will not reflect the same reality that using and
consuming the released supermarket cookbook has in our production
environment.

It is better when using Berkshelf to point to released artifacts - the
released versions of these cookbooks on supermarket itself, so that we
can ensure we're not pulling in potentially hazardous changes that
haven't been released.

Further, it is not clear why certain dependencies are declared here,
such as `ark`, `dmg`, or `7-zip`. Presumably because the entire
dependency graph of the community cookbooks was resolved with these
entries. That is superfluous as Berkshelf will get the community
versions.